토큰 검사 유효한 경우에만 게시글 작성, 삭제,  수정 가능

+++ src/main/java/com/example/springprac2jwt/service/PostService.java
@@ -3,23 +3,44 @@ package com.example.springprac2jwt.service;
 import com.example.springprac2jwt.dto.PostRequestDto;
 import com.example.springprac2jwt.dto.ResponseDto;
 import com.example.springprac2jwt.entity.Post;
+import com.example.springprac2jwt.entity.User;
+import com.example.springprac2jwt.jwt.JwtUtil;
 import com.example.springprac2jwt.repository.PostRepository;
+import com.example.springprac2jwt.repository.UserRepository;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
 
     @Transactional
-    public ResponseDto<?> createPost(PostRequestDto postRequestDto) {
-        Post post = new Post(postRequestDto);
+    public ResponseDto<?> createPost(PostRequestDto postRequestDto, HttpServletRequest request) {
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
+        if(token == null) return ResponseDto.set(false, 401, "토큰이 존재하지 않음");
+        if (jwtUtil.validateToken(token)) {
+            // 토큰에서 사용자 정보 가져오기
+            claims = jwtUtil.getUserInfoFromToken(token);
+        } else {
+            return ResponseDto.set(false, 403, "토큰 에러");
+        }
+        // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
+        User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+        );
+        Post post = new Post(postRequestDto, user);
         postRepository.save(post);
         return ResponseDto.setSuccess(post);
+
     }
 
     @Transactional(readOnly = true)
@@ -33,18 +54,44 @@ public class PostService {
     }
 
     @Transactional
-    public ResponseDto<?> updatePost(Long id, PostRequestDto requestDto) {
+    public ResponseDto<?> updatePost(Long id, PostRequestDto postRequestDto, HttpServletRequest request) {
         Post post = getPostIfExists(id);
-        checkPassword(post, requestDto.getPassword());
-        post.update(requestDto);
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
+        if(token == null) return ResponseDto.set(false, 401, "토큰이 존재하지 않음");
+        if (jwtUtil.validateToken(token)) {
+            // 토큰에서 사용자 정보 가져오기
+            claims = jwtUtil.getUserInfoFromToken(token);
+        } else {
+            return ResponseDto.set(false, 403, "토큰 에러");
+        }
+        // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
+        User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+        );
+
+        if (post.getUser().getUsername().equals(claims.getSubject())) {
+            post.update(postRequestDto);
+        } else return ResponseDto.set(false, 403, "수정할 권한이 없음");
         return ResponseDto.setSuccess(post);
     }
 
     @Transactional
-    public ResponseDto<?> deletePost(Long id, PostRequestDto requestDto) {
+    public ResponseDto<?> deletePost(Long id, HttpServletRequest request) {
         Post post = getPostIfExists(id);
-        checkPassword(post, requestDto.getPassword());
-        postRepository.deleteById(id);
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
+        if(token == null) return ResponseDto.set(false, 401, "토큰이 존재하지 않음");
+        if (jwtUtil.validateToken(token)) {
+            // 토큰에서 사용자 정보 가져오기
+            claims = jwtUtil.getUserInfoFromToken(token);
+        } else {
+            return ResponseDto.set(false, 403, "토큰 에러");
+        }
+        // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
+        if (post.getUser().getUsername().equals(claims.getSubject())) {
+            postRepository.deleteById(id);
+        } else return ResponseDto.set(false, 403, "삭제할 수 없습니다.");
         return ResponseDto.setSuccess(post);
     }
 
@@ -55,10 +102,5 @@ public class PostService {
                 () -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
     }
 
-    private void checkPassword(Post post, String password) {
-        if (!post.getPassword().equals(password)) {
-            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
-        }
-    }
 
 }
